Refactor: extract shared embed out of the entry point
The shared `universalEmbed` was exported from src/index.ts, forcing every
command and listener to import the bootstrap module just to reuse an embed.
Move it to its own utils/embed.ts and repoint all importers. Also modernize
index.ts: use the ActivityType enum instead of the magic number 3, drop the
dead db import comment, and remove the now-unnecessary EmbedBuilder import.
