+++ src/index.ts
@@ -1,8 +1,7 @@
 import { SapphireClient, Logger, LogLevel } from '@sapphire/framework';
-import { EmbedBuilder, GatewayIntentBits } from 'discord.js';
+import { ActivityType, GatewayIntentBits } from 'discord.js';
 import { BOT_TOKEN } from './utils/config';
 import { version } from '../package.json';
-// import './utils/db';
 
 const logger = new Logger(LogLevel.Info);
 
@@ -17,14 +16,9 @@ const client = new SapphireClient({
     GatewayIntentBits.DirectMessages,
   ],
   presence: {
-    activities: [{
-      name: `/help | v${version}`,
-      type: 3
-    }]
-  }
-})
+    activities: [{ name: `/tag • v${version}`, type: ActivityType.Watching }],
+  },
+});
 
-logger.info('Attempting to connect to discord client...')
-client.login(BOT_TOKEN)
-
-export default (new EmbedBuilder().setFooter({ text: 'Arduino Bot • GPL-3.0 • /tag' }).setColor('#dc5b05').toJSON())
+logger.info('Attempting to connect to discord client...');
+void client.login(BOT_TOKEN);
